Stub out db data key rotation
